refactor: Cleanups for benchmark
